+++ fuzzing/src/interface.hpp
@@ -4,13 +4,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021 Datadog, Inc.
 
-#ifndef DDWAF_FUZZER_UTILS_H
-#define DDWAF_FUZZER_UTILS_H
+#pragma once
 
-#include <cstdlib>
 #include <ddwaf.h>
 
 ddwaf_handle init_waf();
 void run_waf(ddwaf_handle handle, ddwaf_object args, size_t timeLeftInUs);
-
-#endif
