+++ fuzzing/src/object_builder.hpp
@@ -4,13 +4,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021 Datadog, Inc.
 
-#ifndef DDWAF_FUZZER_OBJECT_BUILDER_H
-#define DDWAF_FUZZER_OBJECT_BUILDER_H
-
-#include <cstdlib>
+#pragma once
 #include <ddwaf.h>
 
 ddwaf_object build_object(
     const uint8_t *bytes, size_t size, bool verbose, bool fuzzTimeout, size_t *timeLeftInMs);
 
-#endif
